ulysse-common: overlay: Show "System navigation" setting on AOSP
This CL makes "System navigation" setting item available in the system
settings on AOSP coral/flame devices.

Since anyone is already able to do this by manually tweaking this
framework config then building a custom image, there should be no
strong reason to not make it available by default.

Doing this makes it easier for us to work on bugs and issues that are
reported with this setting on.

Bug: 139806621
Test: Manually made sure that the settings is available as follows
  1. lunch aosp_coral-userdebug && make -j
  2. Flash the image
  3. Open the system settings
  4. System -> Gestures
  5. Confirm "System navigation" is shown
  6. Tap "System nagivation"
  7. Tap "Gesture nagivation"
  8. Confirm that gesture nagivation is working
Change-Id: I11c18871d0bb52e6648b4af2a00a6ef4a54cd041